requirements.txt
# Cryptocurrency trading
python-binance
twisted >= 18.4.0
ccxt
tulipy

## Services
# Reddit
praw

# Telegram
python-telegram-bot

# Twitter
Python-Twitter

# Discord
discord.py

# Web
jinja2
flask

## Tools
# Analysis tools
newspaper3k
pytrends
vaderSentiment

# ML tools
tensorflow
tensorflow-tensorboard

## Others
psutil
colorlog
cryptography >= 2.2.1
requests
